# Optimize `ElementRestriction::FillI` and `ElementRestriction::FillJAndData`.